object click and hold.
when you click an object it will turn a different color indicate it has been click.
when you hold an object it will change a color, showing it be holding and when you stop the color will change.
these scripts are togglable and you have to manually apply them to the gameobject for what u want the player to do.